teacher

+++ src/main/java/ecourse/model/TeacherRepository.java
@@ -2,6 +2,6 @@ package ecourse.model;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface TeacherRepository extends CrudRepository<Teacher, Integer> {
-    
+public interface TeacherRepository extends CrudRepository<Teacher, Short> {
+
 }
